requirements.txt
django-jazzmin
Django
pandas
openpyxl
Pillow
dj-database-url
gunicorn
django-import-export
psycopg2 